requirements.txt
setuptools>=65.0.0
wheel>=0.38.0
rank-bm25==0.2.2
nltk==3.9.1
spacy==3.7.5
sentence-transformers==3.0.1
torch==2.5.1
transformers==4.45.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
pyyaml==6.0.2
tqdm==4.66.5
fastapi==0.115.0
uvicorn[standard]==0.32.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl